snows_recon/requirements.txt
Linux tools:
whois
dnsrecon
dnsenum
subfinder
gospider
nikto
gobuster

Python libraries:
queue
subprocess
sys
requests
threading
pyfiglet
dns.resolver
bs4
